requirements.txt
aiogram==3.0.0b7
sqlalchemy[asyncio]==1.4.35
python-dotenv==0.19.2
aiohttp==3.8.4
aiosqlite==0.19.0  # Добавьте эту строку
python-dateutil==2.8.2